Keyframe adjustments, Added border-radius
Adjusted top positions in the @keyframes 'to' statement. Since new
entries were made to the database, the timing of the animation was off,
so adjusting the position of the scroll was nessesarry. This does open a
new issue with this method though. And brings up a great question - Can
I use jQuery to dynamically control this? Sure you can, but the question
is, "Can I?" Looks like a lot more jQuery API Doc reading is in my
future.